tg-{revert,update}.sh: quiet confusing auto stash messages
When either tg update or tg revert performs an auto stash prior to
making changes, some odd warnings (especially with tg revert) can
be emitted by the tg tag command used to do the auto stash.

Suppress these harmless warnings in this case as otherwise they
appear to be related to the tg revert / tg update command which
they actually are not thus leading to confusion.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>