tg-summary.sh: default to --with-deps when sensible
It makes much more sense to default to --with-deps when it
makes sense (considering the other options given).

Provide a new --without-deps option to revert to the previous
behavior.

For example this:

  tg summary @

now works to give a full summary of the TopGit-controlled branch
that HEAD refers to and all of its (recursively determined)
dependencies.

Prevviously this would have just shown one line which is not
at all intuitive.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>