tg-update.sh: allow multiple branch names
This is similar to using -a with multiple patterns except that each
name must be a valid TopGit-controlled branch name (or HEAD or @
assuming HEAD is a symref to a TopGit-controlled branch).

Multiple branch names work the same as multiple patterns in that
a failure updating a branch prior to the last that does NOT occur
in a recursive update fails to mention the need to update any
remaining branches when re-running the tg update command.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>