tg-tag.sh: allow "anonymous" stashes
These work just like "--stash" except they use "--anonymous" instead
and store the result in the bare ref "TG_STASH".

They are not expected to be persistent or particularly visible to
end users or pushable.

Rather the anonymous stash provides an internal mechanism to quietly
save a stash even if the user asked for "--no-stash" so that they
can actually recover.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>