tg-remote.sh: handle new bases location better
Properly handle either all old or all new top bases setup.

This requires special handling for the new location due to naming
and the fact it's located under heads.

Also the warning and removal of old "push" specs is reverted to
the hard-coded old location since the new code would never create
those in the first place.

This is not enough to support a remote with its bases in a
different location that the local repository, but both locations
are now fully supported provided local and remote both use the
same location.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>