tg-depend.sh: expunge use of tsort
Use of tsort to detect dependency loops is unreliable since
tsort is not required to "fail" if loops are present or output
anything particular that's guaranteed to be machine readable.

Instead use the same technique that the pre-commit.sh hook
does and just call recurse_deps to see if the newly added
dependency has a path to the current head.

This makes the command work as expected and eliminates a rather
unexpected (and now gone) dependency on the tsort command.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>