tg.sh: handle remote bases in new location
Different handling is needed for remote bases in the new
location compared to the old one as the leading "heads/"
must be stripped when using the new location.

With this change all commands should now operate correctly
in either the old or new location but only tg create (using
its -r name option) and tg revert (always) handle a mix of
old and new bases locations.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>