test-lib-main.sh: with --verbose-log allow --verbose-only=... and -x
If --verbose-log is specified then the standard out and standard error
of the test scripts passed to test_expect_success... are captured to a log
file and do not disrupt any $HARNESS_ACTIVE parsing.

There's no reason then not to allow --verbose-only=... and/or --xtrace
under these circumstances therefore make it so.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>