test-lib-main.sh: add test_plan function
It's nice to be able to output a fixed test plan count in advance
so TAP harnesses can provide a nice progress display.

Make that possible by adding a new test_plan function.

As a bonus it also provides an easier way to pass a skip message
when outputting a "1..0 # SKIP" plan line.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>