tg-create.sh: remove mention of --continue
The `tg create` comand doesn't have its own `--continue` operation
anymore, it just runs `tg update` under-the-hood as the last step
and so makes use of the `tg update --continue` command instead when
necessary.

Remove all mention of and related logic for the `tg create [--continue]`
command to reflect this.

Also make sure that "@" is understood as "HEAD" where appropriate.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>