tg-create.sh: preserve pre-loaded message during update
The tg create command may conveniently pre-load a commit message
based on the contents of the .topmsg file for the created branch.

Make sure this is preserved across any needed tg update call by
renaming it just before invoking the tg update code and having the
tg update code restore it just before the update completes.

This restores the behavior that was available before tg create
started using the tg update code to do the base merge for multiple
dependencies.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>