tg-create.sh: only preload "[PATCH]" commit msgs
When creating a new topic branch (and NOT using --no-commit) the
first thing likely to be done after the branch is created is to
commit the initial patch.

Since the message for this patch will likely closely resemble the
message already stored in .topmsg, that message (with the headers
suitably stripped) is "preloaded" into the "MERGE_MSG" file so
that it will automatically show up for convenience when the first
commit is made.

Howevver, if the TopGit branch being created is a base, root,
stage or release (or actually any non-patch) branch then there
will probably NOT be a follow-up commit with the patch contents
because such a branch generally does not contain any changes of
its own.

Therefore avoid pre-loading the "MERGE_MSG" unless the .topmsg
"Subject:" contains "[PATCH]".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>