tg-create.sh: abort early if initial checkout would fail
As part of branch creation, the first dependency (if there is one
and it has a tree different than HEAD's) needs to be checked out.

Abort early before collecting the .topmsg if that checkout would
fail for some reason.

Also add the same capability to the "ensure_clean_tree" function
if it's passed an argument.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>