test-lib*.sh: set GIT_CEILING_DIRECTORIES
Set, export and cache GIT_CEILING_DIRECTORIES to make sure
that running tests cannot accidentally find the git directory
of the actual Git repository containing the sources.

Without this and without using a --root option, if the
TEST_NO_CREATE_REPO variable is set to a non-empty value Git
would happily find the repository containing the sources.

That could be very bad indeed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>