tg.sh: st[atus] --help
Make "tg status -h" and "tg status --help" work like it's supposed
to just like all the other commands.

Move the usage into the Makefile so that it can be in just one
single place instead of three since it's handled specially.

And provide an "st" alias for "status" at the same time because,
of all the commands, "tg status" is likely to be run a LOT and
since it never does the TopGit mergesetup stuff it's also perfectly
safe to use on ANY repository and is in some ways more helpful than
"git status" and is now quicker to type too.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>