Makefile: support V=1
When the make variable V is set to anything other than the
empty string or 0, show the actual commands instead of the
single summary line.

This makes it much easier to debug issues when something's
not working right.  The default "quiet" functionality
can be debugged by doing "make AT=" to force its actions
to become visible.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>