tg status: show upstream info and implement verbose
Change the parenthetical information to use '['...']' instead
of '('...')' since '(' and ')' are allowed in ref names but
'[' is not.

Strip a leading "refs/heads/" off the shown HEAD symbolic ref
by default.

Show the upstream ref name (if there is one) with its leading
"refs/remotes/" stripped by default.

Also show how far ahead/behind it is when it's not the same
commit.

With a single --verbose (or -v) also now include a short dirty
(but not untracked) file status listing.  This also causes
non file status lines to be prefixed with "## ".

With two or more -v (or --verbose) options, show full ref names
and unabbreviated hash values as well.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>