tg.sh: verify_topgit_branch correctly for 'HEAD' on checked out base
If HEAD is a symbolic ref to a TopGit top-bases branch then omitting
the branch name for most commands or explicitly supplying either "HEAD"
or "@" should automatically use the TopGit branch corresponding to
the checked out TopGit base, but it didn't.

The logic was there to support explicitly passing in a fully-qualified
top-bases ref, but not if such was obtained by resolving HEAD's symref.

Fix this so that it works properly in this case especially since with
the new top-bases location it's particularly easy to checkout a base
branch to manipulate it.  So this really does need to work properly.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>