tg-summary.sh: include new '*' "needs merge" column
When showing long branch status (the default mode) either for
all dependencies or using the "--with-deps" option, indicate
which of the displayed branches are contributing to the
Delinquency (err, um, the 'D' status anyway) of any of its
direct or indirect dependents (i.e. those branches that
depend on it either directly or indirectly).

This information is extremely helpful when looking at a summary
output with a bunch of 'D' status lines.  Previously there was
no clue which of the other branch(es) contained unmerged commits
that were causing such status displays.

Now there is.  A '*' will appear in the last status column (this
is a newly added status column) if the listed branch is one of
the triggering branches of some other branch's 'D' status.

The computation is done in such a way that it's essentially
free if one or more branches are being explicitly listed on
the command line, otherwise it only incurs the additional
cost of a "tg summary --heads" run (which is usually quite small).

The computation is suppressed when not showing all branches or
using the, possibly implied,  "--with-deps" option because the
results could be incomplete and therefore misleading.  The
column will just always be empty.

This also changes the status columns to separate themselves
from the branch name with a space instead of a tab to avoid
unwanted horizontal space from showing up in the output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>