tg-migrate-bases.sh: handle HEAD symrefs
If the HEAD symref points to a ref being migrated, migrate
the symref too.

As a special case, if run in a linked working tree then
the main working tree's HEAD symref will also be checked
and migrated if needed.

In general, the symrefs of the main and all worktree
directories should be processed, but that's a future
update.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>