tg.sh: quote problem alternate directory paths
Since v2.11.1 Git allows paths in GIT_ALTERNATE_OBJECT_DIRECTORIES to
be double-quoted using C-style escapes.

If running Git v2.11.1 or later support this when the temporary
directory path contains problem characters (i.e. ":" or ";") and
go ahead and create a temporary alternate object store in that case.

This avoids tree turds for such repositories if Git v2.11.1 is
available.

This would hardly be worth doing except that an alternate object store
is required for "tg" to work (even using "read-only" operations) on
read-only Git repositories.  The thing is though, if the Git repository
is read-only then the path to the alternate object store will only
have a problem character in it if the temporary directory location does
which is even less likely than the repository itself.

However it's really only a tiny bit more effort and it will then
allow read-only Git repositories to work with "tg" when using Git v2.11.1
or later and the path to the temporary directory in "$TMPDIR" contains
one of the problem characters.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>