tg-create.sh: create bases without .top* files
When creating the initial base for a new TopGit branch, insert an
additional commit just in front of the new base target to remove
any .topmsg and/or .topdeps files if either is present.

This keeps .topdeps and .topmsg files out of the bases (where they
really do not belong) and also makes a log --first-parent -p walk
of the TopGit branch look a lot better when it gets to .topmsg and
.topdeps files.

Now they will be shown as a removal (if the base was another TopGit
branch) followed by addition of the new versions for the new branch.

This is so so so much easier to read and if combined with the -D
(--irreversible-delete) option it's even better.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>