tg-update.sh: after simple merge failure keep simple merges disabled
If a simple merge fails while attempting to merge in any one of the
dependencies that needs updating into the base, simple merges are
already automatically suppressed for the rest of the dependencies
needing to be updated, but only for the current `tg update` session.

That did not extend outside the current session nor, if conflict
resolution was required, did it survive past the conflict resolution
(even though it should have).

Fix this by remembering the simple-merge-suppression state across
"conflict resolution" + "tg update --continue" sequences and also
continue to suppress simple merging when merging in the remote
branch (if any, itself causing simple merges to be suppressed
if a conflict occurs there and they were not already suppressed)
and when merging the updated base (plus remote branch if any) into
the branch itself.

This guarantees that once fallback to recursive logic happens
when updating a single branch that simple merge logic is disabled
for that branch until it's been successfully updated thus preventing
simple merge "accidents" that can be caused by only using a single
"real" base rather than the possibly synthesized base that recursive
merging may end up creating and using (such as when git merge-base
--all produces more than one result).

Note that once the affected branch has been updated, simple merges
are automatically re-enabled when moving on to the next branch.

Also note that disabling "simple" merges in no way disables
"trivial" merges that do not actually require any "automatic"
merging steps.

The previous "tg update" behavior caught most of the issues and this
change plugs the few remaining holes.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>