tg-update.sh: improve .topdeps and .topmsg merge handling
There are two problems that need to be corrected:

 a) The base(s) should not contain any .topdeps or .topmsg files.
    They are not relevant to the base and only create confusion and
    just do not belong in the base.  They have been getting there
    when a TopGit branch dependency is updated and then merged into
    the dependent's base or by tg create when it first creates a new
    base (a previous commit corrected that).  This must stop.

 b) When a remote branch updates its .topdeps and/or .topmsg files
    the changes are just silently ignored in favor of "ours" when
    the remote branch (not its base) is merged into the local branch.
    This makes it hard to share changes and needs correction.

There are therefore four different cases and merge startegies needed:

 1) "--remove" any present .topdeps and .topmsg files.
    When updating a base the result should not have any .topdeps or
    .topmsg files in it regardless of what's in any of the heads
    participating in the update of the base.

 2) "--theirs" always take theirs during a merge conflict.
    When a remote branch has been updated and needs to be merged
    in, it's merged into the result of updating the base instead
    of the other way around and since the base should not have
    any .topdeps or .topmsg files in it, using --theirs brings
    in the remote branch (NOT its base) .topdeps and .topmsg
    files for merging.

  3) "--ours" always takes ours during a merge conflict.
     This is how the attributes are configured and this is
     what happens by default so if something with a .topdeps
     and/or .topmsg file is manually cherry-picked and/or
     merged into a branch there will be no conflicts and
     the original .topdeps and .topmsg files will be kept.
     If there is no remote branch being merged in this
     prevents anything that might be accidently present
     in the base from being picked up when the branch is
     updated with its base.

  4) "--merge" merges the files together.
     When updating a branch with remote branch updates
     this allows remote branch updates to the .topdeps
     and/or .topmsg file(s) to be picked up.  Otherwise
     they would be silently ignored and essentially lost.

The two merge functions (index-only and in-tree) now take
optional "--remove", "--theirs" and "--merge" options.
(There's no need for "--ours" since it's the default.)

All the places the functions are called are updated to
pass the proper option depending on what's being merged
in.

As a result .topdeps and .topmsg files should now be kept
out of any TopGit branch bases while allowing updates to
them to be picked up from remote branches.

In order to make this work properly for in-tree merging, the
merge-recursive driver is run directly and then the index fixed
up to guarantee the correct .topfiles end up in the final commit.

As a result, it's no longer strictly necessary to have an entry
in the .git/info/attributes file for .topdeps and .topmsg nor
the corresponding .git/config entry for the "always keep ours"
merge driver.  However those remain in place for use by
non-tg-update mergers so they can practice safe tg merging.

And since all this work is being done on the merging control code,
take the time to make sure that conflict markers get meaningful
names rather than just an unhelpful bare hash name while making
the other code changes.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>