tg--index-merge-one-file.sh: be more or less --aggressive
Up until now, index-only merges have always been passing the
--aggressive option to "git read-tree -m" in order to handle
extra cases automatically internal to Git thereby being a bit
more efficient.

If a .topdeps or .topmsg conflict arises the merge-index helper
would always resolve it in favor of "--ours".  Since, until
recently, bases were pretty much guaranteed to have a .topmsg
and .topdeps file that differed from the branch itself this
always tended to result in the correct "--ours" behavior even
when passing "--aggressive" to "git read-tree -m" as none of
the special "--aggressive" cases would ever end up applying to
either .topdeps or .topmsg.

This has now changed.

A base branch will now likely NOT have any .topdeps or .topmsg
files in it (as it should be), but when merged into the branch
itself using "--aggressive" this could trigger the "one side leaves
it untouchec and the other removes it so remove it" resolution for
either or both of .topdeps and .topmsg incorrectly removing them
from the branch itself when the base has been updated and is
merged in.

This was bad.

Correct this by only passing the "--aggressive" option to the
"git read-tree -m" command when using the "--merge" or "--remove"
modes for .topdeps and .topmsg.  In the case of "--merge" the
removal is okay if that's the result of the merge and in the
"--remove" case it doesn't matter because the files will end
up always being removed anyway.

Now though the "tg--index-merge-one-file" script must correctly
handle all of the "--aggressive" cases that were previously
being handled by "git read-tree -m --aggressive" for both TopGit
.topfile files (respecting the desired "--ours" or "--theirs")
and for non-TopGit files (doing the proper "--aggressive" thing).

For the most part the script changes are simple and straightforward;
the trick is to switch from --cacheinfo to --index-info when running
"git update-index" in order to make sure removals are handled properly.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>