tg-update.sh: warn when a new non-existing dependency gets merged
If a remote branch adds a dependency to a .topdeps file it will now
be merged in when "tg update" is run (if the remote is set properly).

When that happens and no such local branch exists (or was able to
be automatically created from a same-named TopGit remote branch)
spew a suitable warning message to the user.

New non-TopGit remote branches are deliberately not automatically
set up as local branches since they will not be automatically updated
by TopGit.  The user must manually keep them updated (if desired)
and to do so the user must know they exist; therefore never automatically
create local versions of new dependencies that have an existing
non-TopGit remote branch by that name but do warn profusely about them
and even hint about the correct ref to use to set them up if there
is one.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>