pre-commit.sh: limit hook's auto_create_local_remote activity
The pre-commit hook has the ability to automatically create a
local version of a remote TopGit branch when a newly added
dependency line is found in a .topdeps file.

Only use this special ability wisely.

Although unlikely, it's conceivable that conflict resolution
for something unrelated to a merge of a remote branch update
into the local branch could result (after manual user
intervention to resolve the conflicts) in a bonus change
to the .topdeps file that introduces a new dependency.

Previously pre-commit would use its special auto_create_local_remote
powers in any such case encountered while a "tg update" is active.

Correct this and limit use of the special auto_create_local_remote
powers to only the very special occasion of the final merge of a
remote branch into the branch itself as this is the one and only
time the auto_create_local_remote powers should be exercised
(and is also the only time "tg update" itself uses them).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>