Replace "tg-migrate-bases.sh: handle HEAD symrefs"
This partially reverts the previous commit that added
tg migrate-bases handling of symrefs.

Always plan to throw one away they say.  Consider it thrown.

Replace the previous code with something more elegant and
that properly handles linked working trees if present.

Any updated symrefs now get their own "symref:" line
that makes clear what's really going on.

Handling updates of symrefs to bases is really an edge case
to begin with and dealing with such in linked working trees
even more so.

However, it seems it's always the edge cases that turn
around and bite you in the *cough*, so here's the support.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>