tg-update.sh: introduce --base mode
The new "tg update" --base mode permits quick and easy updating
of "[BASE]" TopGit branches (typically having been created with
the "tg create --base" command).

Previously it was necessary to manually update the "[BASE]"
branch's base using "git update-ref" which is somewhat error
prone.  Then, when "tg update" was run and the new base merged
into the branch itself the merge message would be generic.

Instead the new "--base" mode of "tg update" safely performs
the "git update-ref" after checking to make sure the update is
a fast-forward (using --force can override) and then immediately
merges that into the branch itself using the supplied message
(or message file) or brings up an editor.  The merge then
contains a nice reason (with minimal effort) for the update if
desired and it's all very quick and easy to do now.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>