tg.sh: fix tg update --abort and clarify --stop
The only current client of the "-f" option to checkout_symref_full
is "tg update --abort".  Unfortunately, due to a typo, the "-f"
feature did not work leading to unexpected errors.

The "-f" option also failed to clear out the "merge active" state;
fix this as well.

Finally, clarify the wording on the "tg status" help for the
"tg update --stop" command to make it clearer that it just
stops the update while keeping all "changes" not just "updates".

Meaning index changes, merge-in-progress changes, etc.  The
previous "updates" perhaps suggested only committed changes
which is wrong since "--stop" is just supposed to stop the
update process itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>